meme

+++ base/views.py
@@ -1,5 +1,7 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate,login,logout
@@ -12,6 +14,10 @@ from .forms import RoomForm
 # Create your views here.
 
 def loginPage(request):
+    page='login'
+    if request.user.is_authenticated:
+        return redirect('home')
+    
     if request.method == 'POST':
         
         username=request.POST.get('username')
@@ -31,12 +37,17 @@ def loginPage(request):
              messages.error(request,'username or password Does Not Exist !')
              
                       
-    context= {}
+    context= {'page': login}
     return render(request, 'base/login_register.html' , context)
 
 def logoutUser(request):
     logout(request)
     return redirect ('home')
+
+def registerPage(request):
+    page='register'
+    return render(request, 'base/login_register.html')
+    
     
     
     
@@ -63,6 +74,8 @@ def room(request,pk):
     context={'room':room}
     return render(request,'base/room.html',context)
 
+@login_required(login_url='/login')
+
 def createRoom(request):
     form=RoomForm()
     if request.method == 'POST':
@@ -76,9 +89,14 @@ def createRoom(request):
     context={'form': form}
     return render (request, 'base/room_form.html',context)
 
+@login_required(login_url='/login')
 def updateRoom(request,pk):
     room=Room.objects.get(id=pk)
     form=RoomForm(instance=room)
+    
+    if request.user != room.host:
+        return HttpResponse('You r NOT have permission')
+    
     if request.method == 'POST':
         form=RoomForm(request.POST,instance=room)
         if form.is_valid():
@@ -89,10 +107,15 @@ def updateRoom(request,pk):
     return render(request,'base/room_form.html',contxet)
 
 
-    
+
+@login_required(login_url='/login')   
 def deleteRoom(request,pk):
+    
     room=Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse('You r NOT have permission')
     if request.method=='POST':
         room.delete()
         return redirect('home')
-    return render(request , 'base/delete.html' ,{'obj':room})    
+    return render(request , 'base/delete.html' ,{'obj':room})  
+
